+++ src/utils/string.js
@@ -1,0 +1,5 @@
+module.exports = {
+    escapeRegExp : (str) => {
+        return str.replace(/[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$]/g, "\\$&");
+    }
+}